finish all

+++ lib/srcs/stdio/ft_fputp.c
@@ -14,8 +14,8 @@
 
 int	ft_fputp(unsigned long long unum, int fd)
 {
-	size_t	ret;
-	size_t	w_ret;
+	int		ret;
+	int		w_ret;
 	char	*str;
 
 	ret = 0;
